+++ src/components/index.ts
@@ -5,7 +5,7 @@ export { default as Divider } from "./Divider/Divider";
 export { default as Dropdown } from "./Dropdown/Dropdown";
 export { default as Input } from "./Input/Input";
 export { default as MenuHeader } from "./MenuHeader/MenuHeader";
-export { default as ModalHashtag } from "./ModalHashtag/ModalHashtag";
+export { default as HashtagChip } from "./HashtagChip/HashtagChip";
 export { default as NoDataContainer } from "./NoDataContainer/NoDataContainer";
 export { default as SearchLayout } from "./SearchLayout/SearchLayout";
 export { default as ServerHeader } from "./ServerHeader/ServerHeader";
